requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-multipart
python-jose[cryptography]
passlib[bcrypt]
ultralytics
transformers
torch
Pillow
numpy
opencv-python
ftfy
regex
tqdm
scikit-learn
